backend/requirements.txt
flask
python-dotenv
psycopg2-binary
ollama
flask_cors
groq